backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
langgraph>=0.2.55
langchain>=0.3.7
langchain-openai>=0.2.10
langchain-community>=0.3.5
litellm
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.24.0
pandas>=2.0.0
Pillow>=10.0.0
gunicorn>=21.2.0
arize>=4.0.0
openinference-instrumentation-langchain>=0.1.0
openinference-instrumentation-openai>=0.1.39
